Move tui-kit-host tests into separate files
